+++ apps/editor-dev/pages/plugins/CenterPlugin.jsx
@@ -1,24 +1,19 @@
 import { View } from "@elf-framework/ui";
 
-function Canvas3D() {
-  return (
-    <canvas
-      style={{
-        position: "absolute",
-        top: 0,
-        left: 0,
-        width: "100%",
-        height: "100%",
-      }}
-    ></canvas>
-  );
-}
-
 export async function CenterPlugin(editor) {
+  const { Canvas3D } = await import("./CanvasView");
+
   editor.registerGroupUI("center-panel", {
     renderView: (
       <View position="relative" height="100%" overflow="hidden">
         <Canvas3D />
+        <button
+          type="button"
+          style={{ position: "absolute" }}
+          onClick={() => console.log(editor.configs.get("yellow"))}
+        >
+          Hello
+        </button>
       </View>
     ),
   });