+++ packages/base-editor/src/components/editor/BaseEditor.jsx
@@ -1,10 +1,15 @@
-import { classnames } from "@elf-framework/sapa";
+import { classnames, useSelf } from "@elf-framework/sapa";
 
 import { Editor, useEditor } from "./Editor";
 
 export class BaseEditor extends Editor {
   template() {
     const editor = useEditor();
+
+    useSelf("editor.plugin.activated", () => {
+      this.refresh();
+    });
+
     return (
       <div
         class={classnames("elf--base-editor", {
@@ -13,7 +18,7 @@ export class BaseEditor extends Editor {
         })}
       >
         {/* 렌더링 영역 중간에 보여지는 UI 는 renderView 로 지정한다. */}
-        {editor.getUIList("renderView")}
+        {editor.isPluginActivated ? editor.getUIList("renderView") : undefined}
       </div>
     );
   }