+++ packages/base-editor/src/managers/PluginManager.js
@@ -1,3 +1,35 @@
+class EditorPlugin {
+  constructor(editor, callback, options) {
+    this.editor = editor;
+    this.callback = callback;
+    this.options = options;
+    this.isActivated = false;
+    this.ret = null;
+  }
+
+  async initialize() {
+    const ret = await this.callback(this.editor, this.options);
+
+    this.isActivated = true;
+
+    return ret;
+  }
+
+  async activate() {
+    if (this.isActivated) {
+      return this.ret;
+    }
+
+    this.ret = await this.initialize();
+
+    return this.ret;
+  }
+
+  deactivate() {
+    this.isActivated = false;
+  }
+}
+
 export class PluginManager {
   constructor(editorContext) {
     this.editorContext = editorContext;
@@ -5,7 +37,7 @@ export class PluginManager {
   }
 
   registerPlugin(func, options = {}) {
-    this.plugins.push([func, options]);
+    this.plugins.push(new EditorPlugin(this.editorContext, func, options));
   }
 
   /**
@@ -14,9 +46,9 @@ export class PluginManager {
    */
   async initializePlugin() {
     return await Promise.all(
-      this.plugins.map(async ([CreatePluginFunction, options]) => {
+      this.plugins.map(async (plugin) => {
         try {
-          return await CreatePluginFunction(this.editorContext, options);
+          return await plugin.activate();
         } catch (e) {
           console.error(e);
           return undefined;