+++ apps/editor-dev/CHANGELOG.md
@@ -1,5 +1,17 @@
 # editor-dev
 
+## 1.0.3
+
+### Patch Changes
+
+- modify config settings
+
+## 1.0.2
+
+### Patch Changes
+
+- support initialize for editor plugin
+
 ## 1.0.1
 
 ### Patch Changes