+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui-dev
 
+## 1.0.2
+
+### Patch Changes
+
+- fixed updateChildren bug
+
 ## 1.0.1
 
 ### Patch Changes