+++ packages/sapa/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/sapa
 
+## 2.0.12
+
+### Patch Changes
+
+- fixed makeElement for vnode
+
 ## 2.0.10
 
 ### Patch Changes